survey-worker/pyproject.toml
[project]
name = "survey-worker"
version = "0.1.0"
description = "RQ Worker for processing survey answers into PostgreSQL"
requires-python = ">=3.11"
dependencies = [
    "redis>=5.0.0",
    "rq>=1.16.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
